Allow time.Time types to be converted to string labels
Some bigquery tables use timestamps as part of their schema
and right now should one try to export the data, the resulting
data points in prometheus have "invalid string" instead of string
converted versions of the timestamps. This change would allow
users to convert timestamps to strings and keep the labels as is.